added plugin

+++ platform/plugins/member/database/migrations/2023_10_16_075332_add_status_column.php
@@ -14,8 +14,11 @@ return new class () extends Migration {
 
     public function down(): void
     {
-        Schema::table('teams', function (Blueprint $table): void {
-            $table->dropColumn('status');
-        });
+        if(Schema::hasTable('teams'))
+        {
+            Schema::table('teams', function (Blueprint $table): void {
+                $table->dropColumn('status');
+            });
+        }
     }
 };
